pyproject.toml
[project]
name = "rocketrecap2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "google-genai>=1.36.0",
    "psycopg2-binary>=2.9.10",
    "webvtt-py>=0.5.1",
    "yt-dlp>=2025.9.5",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
]
prod = [
    "gunicorn>=23.0.0",
]
